+++ practice-two/src/components/PricesFilterPopover.jsx
@@ -64,7 +64,6 @@ const PricePopover = ({
                     size="medium"
                     data-min={price.min}
                     data-max={price.max}
-                    style={{ display: 'block' }}
                     onClick={handleSelectPrice}
                   >
                     {price.text}
